Scuffed TP support
- Added TP support and autosolver

I will say the autosolver has every case hard coded so it is entirely possible at least one case is wrong. Also there is a funky bug where it doesn't fully draw a line if you rapidly zoom while it's drawing but who is gonna do that on TP anyways?